hooksを整理
