+++ frontend/src/containers/Channels/Show.tsx
@@ -3,9 +3,9 @@ import { SongItems } from '@/components/SongLists/SongItems'
 import { SongItemsSearch } from '@/components/SongLists/SongItemsSearch'
 import { Videos } from '@/components/SongLists/Videos'
 import { VideosSearch } from '@/components/SongLists/VideosSearch'
+import { useVideos } from '@/hooks/video'
 import { ChannelType } from '@/resources/types'
 import { getData } from '@/utils/api'
-import { useVideos } from '@/utils/hooks'
 import Error from 'next/error'
 import { useState } from 'react'
 import useSWR from 'swr'
